Squashed commit of the following:
commit b28ce11c68a0c5e610d495b9c215a2e2733d4b47
Author: Noah D. Brenowitz <nbren12@gmail.com>
Date:   Thu May 6 06:21:30 2021 -0700

    Random Cleanups (#381)

    * cleanup _request function

    Group the param, header, and path formatting into separate functions

    * refactor credential logic to separate class

    this seemed like a clean separation of concerns, and makes the
    GCSFileSystem more coherent.

    * delete unused auth module

    this does not appear to be referenced anywhere else in the code or
    documentation.

    * refactor _get_file to use shared functions

    * fix errors finding project

    * move GoogleCredentials to credentials.py

    This reduces the size of core.py

commit f53b58f32da8cd2d510d966b0859bac01a854b5d
Author: Oliver Watt-Meyer <oliwm@vulcan.com>
Date:   Wed May 5 06:11:12 2021 -0700

    Retry decorator for _call and _get_file (#387)

    * Move retry logic to decorator defined in utils.py

    * Use retry decorator for _get_file

    * Fix logger import

    * Rename utils.py to retry.py

    * Delete unused FileSender class

    * Move validate_response with func for cleaner retry decorator

    * Clean up tests

    * Restore argument order

    * Format

    * Format for smaller diff

    * Use func.__name__ in error messages
